1월 7일 커밋

+++ src/main/java/com/meomureum/springboot/dto/PlaceDTO.java
@@ -3,13 +3,13 @@ package com.meomureum.springboot.dto;
 import lombok.Data;
 
 @Data
+
 public class PlaceDTO {
-	
 	private String p_code;
 	private String api_code;
 	private String p_place;
 	private String p_category;
-	private double p_lat;
-	private double p_lon;
+	private int p_lat;
+	private int p_lon;
 	private String p_addr;
 }
